backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymysql==1.1.1
cryptography==43.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
bcrypt==3.2.0
sqlalchemy==2.0.35
PyPDF2==3.0.1
pdfplumber==0.11.7
python-docx==1.1.2
requests==2.32.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
openai==1.55.3
PyMuPDF==1.26.1
pymupdf==1.26.1
